https://online.sbis.ru/opendoc.html?guid=8706ba09-11bc-41d8-bdc5-2402073a74c1

+++ Types/_object/merge.ts
@@ -15,12 +15,10 @@ function mergeInner<T, S>(path: Set<T>, target: T, ...sources: S[]): T & S {
                 const overrides = {};
                 Object.keys(source).forEach((key) => {
                     const sourceItem = source[key];
-                    if (isObject(sourceItem)) {
-                        if (target.hasOwnProperty(key) && !path.has(sourceItem)) {
-                            path.add(sourceItem);
-                            overrides[key] = mergeInner(path, target[key], sourceItem);
-                            path.delete(sourceItem);
-                        }
+                    if (isObject(sourceItem) && isObject(target[key]) && !path.has(sourceItem)) {
+                        path.add(sourceItem);
+                        overrides[key] = mergeInner(path, target[key], sourceItem);
+                        path.delete(sourceItem);
                     }
                 });
